+++ src/main/scala/org/podval/tools/backend/BackendExtension.scala
@@ -3,6 +3,7 @@ package org.podval.tools.backend
 import groovy.lang.Closure
 import org.gradle.api.Project
 import org.podval.tools.build.{ScalaBackend, ScalaDependency, ScalaLibrary, Version}
+import org.podval.tools.gradle.Extensions
 import org.podval.tools.nonjvm.NonJvmBackend
 import org.podval.tools.test.framework.Framework
 import javax.inject.Inject
@@ -12,7 +13,7 @@ abstract class BackendExtension @Inject(
   project: Project,
   val getBackend: ScalaBackend,
   val isRunningInIntelliJ: Boolean
-) extends Logging(project):
+) extends WithProject(project):
   final def getName      : String = getBackend.name
   final def getSourceRoot: String = getBackend.sourceRoot
   final def getSuffix    : String = getBackend.artifactNameSuffix(getScalaBinaryVersion)
@@ -89,6 +90,22 @@ abstract class BackendExtension @Inject(
     )
 
 object BackendExtension:
+  val name: String = "scalaBackend"
+
+  def create(
+    project: Project,
+    backend: ScalaBackend,
+    isRunningInIntelliJ: Boolean
+  ): BackendExtension = Extensions.create(
+    project,
+    name,
+    classOf[BackendExtension],
+    backend,
+    isRunningInIntelliJ
+  )
+
+  def get(project: Project): BackendExtension = Extensions.getByName(project, name)
+
   private type Configure = Closure[ScalaDependency]
   private def id: Configure = Closure.IDENTITY.asInstanceOf[Configure]
   private def toFunction(configure: Configure): ScalaDependency => ScalaDependency = configure.call