+++ src/main/scala/org/podval/tools/backend/SharedProject.scala
@@ -2,7 +2,7 @@ package org.podval.tools.backend
 
 import org.gradle.api.{Project, Task}
 import org.podval.tools.build.ScalaBackend
-import org.podval.tools.gradle.{Projects, Tasks}
+import org.podval.tools.gradle.Tasks
 
 final class SharedProject(
   project: Project,
@@ -10,14 +10,12 @@ final class SharedProject(
 ) extends SingleProject(
   project
 ):
-  override def apply(): Unit =
-    announce(s"code shared between backends ${ScalaBackend.names(backends)}")
+  override def announcement: String = s"code shared between backends ${ScalaBackend.names(backends)}"
 
+  override def apply(): Unit =
     // Disable all tasks.
     Tasks.disable(project, classOf[Task])
 
-    // Set Scala version from parent.
-    setScalaVersionFromParent(Projects.parent(project).get)
-
-    // Add version-specific sources.
-    project.afterEvaluate((_: Project) => addVersionSpecificSources())
+  override def afterEvaluate(): Unit =
+    ()
+