+++ src/main/scala/org/podval/tools/gradle/Projects.scala
@@ -2,7 +2,6 @@ package org.podval.tools.gradle
 
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.{Plugin, Project}
-import org.gradle.api.internal.project.ProjectStateInternal
 import org.podval.tools.platform.Files
 import java.io.File
 import scala.jdk.CollectionConverters.SetHasAsScala
@@ -24,7 +23,7 @@ object Projects:
 
   // We look up projects by their *directory* names, not by their *project* names,
   // so `Option(project.findProject(name))` does not do it for projects renamed in `settings.gradle` ;)
-  def findSubproject(project: Project, name: String): Option[Project] = project
+  def findSubProject(project: Project, name: String): Option[Project] = project
     .getSubprojects
     .asScala
     .find(projectDirName(_) == name)
@@ -43,11 +42,3 @@ object Projects:
   def setBuildSubDirectory(project: Project, subDirectoryName: String): Unit = 
     val property: DirectoryProperty = buildDirectoryProperty(project)
     property.set(property.get.dir(subDirectoryName))
-
-  def afterEvaluateIfAvailable(project: Project, action: => Unit): Unit =
-    val state: ProjectStateInternal = project.getState.asInstanceOf[ProjectStateInternal]
-    val isAfterEvaluateAvailable: Boolean = state.isUnconfigured || state.isConfiguring
-
-    if isAfterEvaluateAvailable
-    then project.afterEvaluate(_ => action)
-    else action